[python] Use of float32 during evaluation

+++ python/eval_model.py
@@ -1,9 +1,9 @@
+import lasagne
 import numpy as np
+
 from c3d_feature_helper import Feature
 from utils import segment_format
 
-import lasagne
-from lasagne.layers import set_all_param_values
 
 def forward_pass(network, input_data):
     """Forward pass input_data over network
@@ -12,6 +12,7 @@ def forward_pass(network, input_data):
                                                        deterministic=True)
     loc = l_pred_var.eval().reshape((-1, 2))
     return loc, y_pred_var.eval()
+
 
 def nms_detections(dets, score, overlap=0.3):
     """
@@ -24,7 +25,7 @@ def nms_detections(dets, score, overlap=0.3):
 
     Parameters
     ----------
-    dets : ndarray. 
+    dets : ndarray.
         Each row is ['f-init', 'f-end']
     score : 1darray.
         Detection score.
@@ -58,8 +59,9 @@ def nms_detections(dets, score, overlap=0.3):
 
     return dets[pick, :], score[pick]
 
+
 def retrieve_proposals(video_name, l_size, network, T=256, stride=128,
-                       c3d_size=16, c3d_stride=8, pool_type='mean', 
+                       c3d_size=16, c3d_stride=8, pool_type='mean',
                        hdf5_dataset=None, model_prm=None):
     """Retrieve proposals for an input video.
 
@@ -93,14 +95,14 @@ def retrieve_proposals(video_name, l_size, network, T=256, stride=128,
     # Video scanning.
     f_init_array = np.arange(0, l_size - T, stride)
     feat_stack = fobj.read_feat_batch_from_video(video_name, f_init_array,
-                                                 duration=T)
+                                                 duration=T).astype(np.float32)
     if model_prm.startswith('lstm:'):
         user_prm = model_prm.split(':', 1)[1].split(',')
         n_outputs, seq_length, width, depth = user_prm
         feat_stack = feat_stack.reshape(feat_stack.shape[0],
                                         int(seq_length),
                                         feat_stack.shape[1]/int(seq_length))
-        
+
     # Close instance.
     fobj.close_instance()
 
@@ -109,7 +111,7 @@ def retrieve_proposals(video_name, l_size, network, T=256, stride=128,
     n_proposals = score.shape[1]
     n_segments = score.shape[0]
     score = score.flatten()
-    map_array = np.stack((f_init_array, 
+    map_array = np.stack((f_init_array,
                           np.zeros(n_segments))).repeat(n_proposals, axis=-1).T
     proposal = segment_format(map_array + (loc.clip(0, 1) * T),
                               'c2b').astype(int)
